Authorization caching implemented
There was a massive issue with Azure BLOB storage being constantly read, which was caused by _authorization.json blob being read on every authorizable call to the back end (which is nearly all calls).

This commit includes authorization cache feature, which caches the authorization state for 10 minutes. This reduces the amount of reads to _authorization.json by a factor of at leasat 40.