Back-end improvements (autoSyncIds function + more)
Fixed bug with sorting. I am not sure whether this is JavaScript by-design behavior, but sorting of number[] resulted in [10, 11, 12, 9] instead of [9, 10, 11, 12] (actually seems to be by-design). Provided compare function that takes care of sorting numbers correctly.

Implemented autoSyncIds function that can accept syncIds payloads for multiple apps at the same time. This not only reduces the number of calls (which is insignificant in and out of itself) but it makes sure multiple consecutive syncIds calls don't fail due to 429 Too Many Requests.