Adding the v2/news function + testability
This will be used to "push" announcements to Ninja. They will be shown on load and then every two hours while the user is working in the app.

At the same time a minor bug was fixed that affected testability. When body was not present, testability injected {} instead, and this made test not correctly resemble real life scenarios where {} was not automatically injected.

Also, request handlers and validators were fixed to not expect body to be defined before running.