Fixing some bugs identified during testing
Testability didn't properly detect issues with some calls because the storage stub was not behaving exactly as actual storage would (it was returning {} instead of undefined where missing blob was being accessed).

Fixed the stub, the tests, and the functionality that was broken, but not identified as such due to bad tests.