Fixing bug with reporting existing consumption
Due to the structure of new back-end storage, existing consumption must be reported as true even if there are no objects of that type in there. That's because synchronizing an empty repository will not result in app cache containing only {}, in which case `hasConsumption` property must still be true. Otherwise, the front-end will keep asking about "Do you want to synchronize" and synchronization will not really change anything, until at least one object of the specified type is assigned.

To avoid this, I am now returning true when app cache JSON object exists in blob.

For the next iteration of getNext this will all be changed anyway, current behavior is here just to support existing version of front end.