Detection and auto-fixing of invalid logical ranges
Theoretically, somebody can define a range where from is higher than to. AL Language prevents this in app.json, but Ninja didn't do anything about it in .objidconfig.

Now, Ninja detects this problem and gives an error with an option to auto-fix. However, it always makes sure that ranges are valid and for all practical purposes, it will treat these kinds of ranges as valid.

Furthermore, if a range does not use number types for from and to values, it shows an error and excludes such ranges from consideration (until fixed).