Massive refactoring of app monitoring
This is a part of ongoing refactoring of app, manifest, config, and workspace handling in Ninja.

Previously there were a series of unorganized and inefficient functions, classes, statics, etc.

This ongoing refactoring is replacing all that with WorkspaceWatcher class that centralizes all these operations.

In this commit, the old (now obsoleted, and deleted) ObjIdConfigMonitor is replaced with new functionality (and is improved while at that).