Making tree view a bit more abstract
This kind of abstraction is needed for the upcoming App Pool Explorer feature. While Range Explorer is per app and every root node has 1:1 relationship with an app (and any change in app structure directly reflects on tree node structure), the App Pool Explorer has 1:n relationship with apps (and change in app structure may not need to reflect on tree node structure).