Back end: support for assigning and unassigning
This commit includes back-end support for explicit assigning and unassigning of individual IDs. This back-end feature is needed to support the front-end feature for management of unassigned IDs (object IDs in use by the workspace, but not assigned by Ninja) and unused IDs (object IDs assigned by Ninja, but no longer in use by the workspace).